Create functions.php data.php, remove function esc, correct other mistakes
