+++ functions.php
@@ -1,0 +1,8 @@
+<? 
+// функция форматирования числа
+function editNumber($number) {
+	$number = ceil($number);
+	$number = number_format($number, 0, ',', ' ');
+	$number.= ' ₽';
+	return $number;
+}
