+++ templates/index.php
@@ -22,15 +22,15 @@
         <?php foreach ($ad as $lot): ?>
             <li class="lots__item lot">
                 <div class="lot__image">
-                    <img src="<?=esc($lot['url_img']);?>" width="350" height="260" alt="">
+                    <img src="<?=htmlspecialchars($lot['url_img']);?>" width="350" height="260" alt="">
                 </div>
                 <div class="lot__info">
-                    <span class="lot__category"> <?=esc($lot['category']);?> </span>
-                    <h3 class="lot__title"><a class="text-link" href="pages/lot.html"> <?=esc($lot['name']);?> </a></h3>
+                    <span class="lot__category"> <?=$lot['category'];?> </span>
+                    <h3 class="lot__title"><a class="text-link" href="pages/lot.html"> <?=htmlspecialchars($lot['name']);?> </a></h3>
                     <div class="lot__state">
                         <div class="lot__rate">
                             <span class="lot__amount"> Стартовая цена </span>
-                            <span class="lot__cost"> <?=esc(editNumber($lot['price']))?> <b class="rub">р</b></span>
+                            <span class="lot__cost"> <?=htmlspecialchars(editNumber($lot['price']))?> <!--b class="rub">р</b--></span>
                         </div>
                         <div class="lot__timer timer">
                             12:23
