+++ templates/layout.php
@@ -24,7 +24,7 @@
         <nav class="user-menu">
             <?php if ($is_auth): ?>
                 <div class="user-menu__logged">
-                    <p> <?=$user_name ?> </p>
+                    <p> <?=htmlspecialchars($user_name)?> </p>
                     <a class="user-menu__bets" href="pages/my-bets.html">Мои ставки</a>
                     <a class="user-menu__logout" href="#">Выход</a>
                 </div>
